Use new board API in iOS app.

+++ game-engine/lib/xiangqi/internal/agents/random.cc
@@ -16,7 +16,8 @@ using ::xq::Position;
 uint16_t Random::MakeMove(const Board<Piece>& board, Player player) const {
   const std::vector<uint16_t> possible_moves =
       AllPossibleNextMoves(board, player);
-  std::uniform_int_distribution<> distrib(0, possible_moves.size() - 1);
+  std::uniform_int_distribution<> distrib(
+      0, static_cast<int>(possible_moves.size()) - 1);
   std::random_device rd;
   std::mt19937 rand_gen(rd());
   int rand_idx = distrib(rand_gen);
